Update README.md
In this sentiment analysis problem our task is to classify a set of tweet into two categories racist/sexist or  non racist/sexist. Sentiment analysis is one of the many applications of Natural Language Processing. It is a set of methods and techniques used for extracting subjective information from text or speech.